XML tag key text: RelativePathInInclude XML tag value: (0 or false to not forbid relative path up access; anything else to forbid them.), it's on by default Error type: RELATIVE_PATH_IN_INCLUDE Error tag: RPI New function: CheckRelativePathInInclude
Tests:
Except for the second test, kwsRunKWStyleTest, all tests are passing. kwsRunKWStyleTest is failing without my changes which indicates an unrelated problem (visual studio 2012 on windows).

Note:
This branch is based off the one in pull request #8, so it includes 2 features.
Pull request #7 needs to be applied for the line numbers to be calculated correctly or use the LineLength rule before this one in the xml file.